:recycle: [Refactor]: Shadcn/ui로 설치된 컴포넌트도 네이밍 컨벤션을 따르도록 수정
