+++ src/shared/ui/index.ts
@@ -1,0 +1,3 @@
+export { Badge } from "./Badge";
+export { Button } from "./Button";
+export { ScrollArea } from "./ScrollArea";
